feat: reduce binary size of libpact_ffi / verifier_cli [profile.release] strip = true opt-level = "z" codegen-units = 1
we do not use lto=true, as this config is applied at the root level to both the pact_ffi and verifier_cli crates. lto=true inflates the size of the static libraries dramatically. It may be prudent to apply this in the release scripts for the verifier cli to allow for further executable size reductions.

Should we elect for setting these values conditionally or via a seperate profile, such that we dont affect the --release profile for other crates in the workspace?